quickroom-backend/SentimentAnalysis/requirements.txt
functions-framework==3.*
requests==2.25.1
Flask==2.0.1
Werkzeug==2.0.3
requests
google-cloud-bigquery==3.*
firebase-admin==6.0.1